+++ src/RiseSim/ViewModels/Controls/CludeRowViewModel.cs
@@ -23,7 +23,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace RiseSim.ViewModels
+namespace RiseSim.ViewModels.Controls
 {
     internal class CludeRowViewModel : BindableBase
     {
