+++ src/RiseSim/Views/Controls/CharmRow.xaml.cs
@@ -15,7 +15,6 @@
  *    along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 using RiseSim.ViewModels;
-using SimModel.model;
 using System;
 using System.Collections.Generic;
 using System.Linq;
@@ -31,14 +30,14 @@ using System.Windows.Media.Imaging;
 using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace RiseSim.Views
+namespace RiseSim.Views.Controls
 {
     /// <summary>
-    /// EquipSelectRow.xaml の相互作用ロジック
+    /// CharmRow.xaml の相互作用ロジック
     /// </summary>
-    public partial class EquipSelectRow : UserControl
+    public partial class CharmRow : UserControl
     {
-        public EquipSelectRow()
+        public CharmRow()
         {
             InitializeComponent();
         }
