+++ src/RiseSim/Views/Controls/CludeRow.xaml
@@ -1,15 +1,18 @@
-﻿<UserControl x:Class="RiseSim.Views.CharmRow"
+﻿<UserControl x:Class="RiseSim.Views.Controls.CludeRow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:local="clr-namespace:RiseSim.Views"
              mc:Ignorable="d" 
-             d:DesignHeight="20" d:DesignWidth="330">
+             d:DesignHeight="20" d:DesignWidth="450">
+
     <Border BorderThickness="0,0,0,1" BorderBrush="Black">
         <StackPanel Orientation="Horizontal">
+            <TextBlock Width="60" Text="{Binding DispKind.Value}" TextAlignment="Right"/>
             <TextBlock Width="300" Text="{Binding DispName.Value}"/>
-            <Button x:Name="deleteCharmButton" Width="30" Content="削除" Command="{Binding DeleteCharmCommand}"></Button>
+            <TextBlock Width="60" Text="{Binding Status.Value}"></TextBlock>
+            <Button x:Name="deleteCludeButton" Width="30" Content="解除" Command="{Binding DeleteCludeCommand}"></Button>
         </StackPanel>
     </Border>
 </UserControl>
