+++ src/RiseSim/Views/Controls/EquipRow.xaml
@@ -1,9 +1,9 @@
-﻿<UserControl x:Class="RiseSim.Views.EquipRow"
+﻿<UserControl x:Class="RiseSim.Views.Controls.EquipRow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:RiseSim.Views" xmlns:viewmodels="clr-namespace:RiseSim.ViewModels" d:DataContext="{d:DesignInstance Type=viewmodels:EquipRowViewModel}"
+             xmlns:local="clr-namespace:RiseSim.Views" xmlns:viewmodels="clr-namespace:RiseSim.ViewModels.Controls" d:DataContext="{d:DesignInstance Type=viewmodels:EquipRowViewModel}"
              mc:Ignorable="d" 
              d:DesignHeight="20" d:DesignWidth="220">
     <StackPanel Orientation="Horizontal">
