+++ src/RiseSim/Views/Controls/EquipSelectRow.xaml
@@ -1,14 +1,14 @@
-﻿<UserControl x:Class="RiseSim.Views.EquipSelectRow"
+﻿<UserControl x:Class="RiseSim.Views.Controls.EquipSelectRow"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:RiseSim.Views" xmlns:viewmodels="clr-namespace:RiseSim.ViewModels" d:DataContext="{d:DesignInstance Type=viewmodels:EquipSelectRowViewModel}"
+             xmlns:local="clr-namespace:RiseSim.Views" xmlns:viewmodels="clr-namespace:RiseSim.ViewModels.Controls" d:DataContext="{d:DesignInstance Type=viewmodels:EquipSelectRowViewModel}"
              mc:Ignorable="d" 
              d:DesignHeight="27" d:DesignWidth="250">
     <StackPanel Orientation="Horizontal">
-        <TextBlock Width="60" Text="{Binding DispKind.Value}" TextAlignment="Right"/>
-        <ComboBox x:Name="equipCombo" Width="130" ItemsSource="{Binding Equips.Value}" SelectedItem="{Binding SelectedEquip.Value}">
+        <TextBlock Width="60" Text="{Binding DispKind.Value}" TextAlignment="Right" VerticalAlignment="Center"/>
+        <ComboBox x:Name="equipCombo" Height="27" Width="130" ItemsSource="{Binding Equips.Value}" SelectedItem="{Binding SelectedEquip.Value}">
             <ComboBox.ItemTemplate>
                 <DataTemplate>
                     <TextBlock Text="{Binding DispName}" Margin="2.5"/>
