+++ src/RiseSim/Views/Controls/SkillSelector.xaml
@@ -1,4 +1,4 @@
-﻿<UserControl x:Class="RiseSim.Views.SkillSelector"
+﻿<UserControl x:Class="RiseSim.Views.Controls.SkillSelector"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
