+++ src/RiseSim/Views/SubViews/CharmTabView.xaml
@@ -1,0 +1,53 @@
+﻿<UserControl x:Class="RiseSim.Views.SubViews.CharmTabView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RiseSim.Views.SubViews"
+             xmlns:ctrl="clr-namespace:RiseSim.Views.Controls" 
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid  ShowGridLines="True">
+
+        <!-- グリッド設定 -->
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="150"/>
+            <ColumnDefinition />
+        </Grid.ColumnDefinitions>
+
+        <!-- 登録 -->
+        <StackPanel Grid.Row="0" Grid.Column="0">
+            <StackPanel>
+                <ItemsControl ItemsSource="{Binding CharmSkillSelectorVMs.Value}">
+                    <ItemsControl.ItemTemplate>
+                        <DataTemplate>
+                            <ctrl:SkillSelector/>
+                        </DataTemplate>
+                    </ItemsControl.ItemTemplate>
+                </ItemsControl>
+            </StackPanel>
+            <StackPanel Orientation="Horizontal">
+                <TextBlock Width="50" Text="スロット"/>
+                <ComboBox x:Name="charmWeaponSlotsBox" ItemsSource="{Binding SlotMaster.Value}"  SelectedItem="{Binding CharmWeaponSlots.Value}" Width="100" SelectedIndex="0">
+                </ComboBox>
+            </StackPanel>
+            <Button x:Name="charmAddButton" Content="登録する" Command="{Binding AddCharmCommand}" />
+        </StackPanel>
+
+        <!-- 一覧 -->
+        <ScrollViewer Grid.Row="0" Grid.Column="1">
+            <StackPanel>
+                <ItemsControl ItemsSource="{Binding CharmRowVMs.Value}">
+                    <ItemsControl.ItemTemplate>
+                        <DataTemplate>
+                            <ctrl:CharmRow/>
+                        </DataTemplate>
+                    </ItemsControl.ItemTemplate>
+                </ItemsControl>
+            </StackPanel>
+        </ScrollViewer>
+    </Grid>
+</UserControl>
