+++ src/RiseSim/Views/SubViews/CludeTabView.xaml
@@ -1,0 +1,45 @@
+﻿<UserControl x:Class="RiseSim.Views.SubViews.CludeTabView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RiseSim.Views.SubViews"
+             xmlns:ctrl="clr-namespace:RiseSim.Views.Controls"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid  ShowGridLines="True">
+
+        <!-- グリッド設定 -->
+        <Grid.RowDefinitions>
+            <RowDefinition/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition Width="250"/>
+            <ColumnDefinition />
+        </Grid.ColumnDefinitions>
+
+        <!-- 登録 -->
+        <StackPanel Grid.Row="0" Grid.Column="0">
+            <ItemsControl ItemsSource="{Binding EquipSelectRowVMs.Value}">
+                <ItemsControl.ItemTemplate>
+                    <DataTemplate>
+                        <ctrl:EquipSelectRow/>
+                    </DataTemplate>
+                </ItemsControl.ItemTemplate>
+            </ItemsControl>
+        </StackPanel>
+
+        <!-- 一覧 -->
+        <ScrollViewer Grid.Row="0" Grid.Column="1">
+            <StackPanel>
+                <ItemsControl ItemsSource="{Binding CludeRowVMs.Value}">
+                    <ItemsControl.ItemTemplate>
+                        <DataTemplate>
+                            <ctrl:CludeRow/>
+                        </DataTemplate>
+                    </ItemsControl.ItemTemplate>
+                </ItemsControl>
+            </StackPanel>
+        </ScrollViewer>
+    </Grid>
+</UserControl>
