+++ src/RiseSim/Views/SubViews/LicenseTabView.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl x:Class="RiseSim.Views.SubViews.LicenseTabView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RiseSim.Views.SubViews"
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid ShowGridLines="True">
+        <Grid.RowDefinitions>
+            <RowDefinition Height="*"/>
+        </Grid.RowDefinitions>
+        <Grid.ColumnDefinitions>
+            <ColumnDefinition />
+            <ColumnDefinition Width="220" />
+        </Grid.ColumnDefinitions>
+        <TextBox x:Name="license" TextWrapping="Wrap" Grid.Row="0" Grid.Column="0" Text="{Binding License.Value}" IsReadOnly="True">
+        </TextBox>
+        <TextBox x:Name="whatIsThis" TextWrapping="Wrap" Grid.Row="0" Grid.Column="1" Text="{Binding WhatIsLicense.Value}" IsReadOnly="True">
+        </TextBox>
+    </Grid>
+</UserControl>
