+++ src/RiseSim/Views/SubViews/LicenseTabView.xaml.cs
@@ -1,0 +1,28 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using System.Windows;
+using System.Windows.Controls;
+using System.Windows.Data;
+using System.Windows.Documents;
+using System.Windows.Input;
+using System.Windows.Media;
+using System.Windows.Media.Imaging;
+using System.Windows.Navigation;
+using System.Windows.Shapes;
+
+namespace RiseSim.Views.SubViews
+{
+    /// <summary>
+    /// LicenseTabView.xaml の相互作用ロジック
+    /// </summary>
+    public partial class LicenseTabView : UserControl
+    {
+        public LicenseTabView()
+        {
+            InitializeComponent();
+        }
+    }
+}
