+++ src/RiseSim/Views/SubViews/MySetTabView.xaml
@@ -1,0 +1,64 @@
+﻿<UserControl x:Class="RiseSim.Views.SubViews.MySetTabView"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:RiseSim.Views.SubViews"
+             xmlns:ctrl="clr-namespace:RiseSim.Views.Controls" 
+             mc:Ignorable="d" 
+             d:DesignHeight="450" d:DesignWidth="800">
+    <Grid>
+        <Grid>
+
+            <!-- グリッド設定 -->
+            <Grid.RowDefinitions>
+                <RowDefinition/>
+            </Grid.RowDefinitions>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition />
+                <ColumnDefinition Width="220" />
+            </Grid.ColumnDefinitions>
+
+            <!-- マイセット一覧 -->
+            <DataGrid x:Name="mySetGrid"  Grid.Row="0" Grid.Column="0" ItemsSource="{Binding MySetList.Value}" IsReadOnly="true" AutoGenerateColumns="False" SelectedItem="{Binding MyDetailSet.Value}">
+                <DataGrid.Columns>
+                    <DataGridTextColumn Header="防御" Binding="{Binding Maxdef}" />
+                    <DataGridTextColumn Header="武器スロ" Binding="{Binding WeaponSlotDisp}" />
+                    <DataGridTextColumn Header="発動スキル" Binding="{Binding SkillsDisp}" />
+                    <DataGridTextColumn Header="火" Binding="{Binding Fire}" />
+                    <DataGridTextColumn Header="水" Binding="{Binding Water}" />
+                    <DataGridTextColumn Header="雷" Binding="{Binding Thunder}" />
+                    <DataGridTextColumn Header="氷" Binding="{Binding Ice}" />
+                    <DataGridTextColumn Header="龍" Binding="{Binding Dragon}" />
+                    <DataGridTextColumn Header="頭" Binding="{Binding HeadName}" />
+                    <DataGridTextColumn Header="胴" Binding="{Binding BodyName}" />
+                    <DataGridTextColumn Header="腕" Binding="{Binding ArmName}" />
+                    <DataGridTextColumn Header="腰" Binding="{Binding WaistName}" />
+                    <DataGridTextColumn Header="足" Binding="{Binding LegName}" />
+                    <DataGridTextColumn Header="護石" Binding="{Binding CharmNameDisp}" />
+                    <DataGridTextColumn Header="装飾品" Binding="{Binding DecoNameCSV}" />
+                </DataGrid.Columns>
+            </DataGrid>
+
+            <!-- 装備詳細出力 -->
+            <StackPanel Grid.Row="0" Grid.Column="1">
+                <Button x:Name="mySetDeleteButton" Content="この装備をマイセットから削除する" Command="{Binding DeleteMySetCommand}"/>
+                <Button x:Name="mySetToConditionButton" Content="この装備のスキルを検索条件に入力する" Command="{Binding InputMySetConditionCommand}"/>
+                <StackPanel Orientation="Horizontal">
+                    <TextBlock Width="60" Text="武器スロ：" TextAlignment="Right"/>
+                    <TextBlock Width="100" Text="{Binding MyDetailSet.Value.WeaponSlotDisp}"/>
+                </StackPanel>
+                <StackPanel>
+                    <ItemsControl ItemsSource="{Binding MyEquipRowVMs.Value}">
+                        <ItemsControl.ItemTemplate>
+                            <DataTemplate>
+                                <ctrl:EquipRow/>
+                            </DataTemplate>
+                        </ItemsControl.ItemTemplate>
+                    </ItemsControl>
+                </StackPanel>
+            </StackPanel>
+
+        </Grid>
+    </Grid>
+</UserControl>
